# backend/src/routes/api/v1/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Webflow Feedback Tool API",
    "version": "1.0.0",
    "description": "REST API for programmatic access to feedback data, project management, and webhook configuration. Available to Premium and Agency tier subscribers.",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://api.yourapp.com/api/v1",
      "description": "Production"
    },
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "BearerApiKey": []
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerApiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key obtained from POST /api/v1/auth/keys. Use as: Authorization: Bearer wfapi_..."
      },
      "BearerJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token for key management endpoints"
      }
    },
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["INVALID_API_KEY", "INSUFFICIENT_PERMISSIONS", "RESOURCE_NOT_FOUND", "VALIDATION_ERROR", "RATE_LIMIT_EXCEEDED", "INTERNAL_ERROR"]
              },
              "message": { "type": "string" },
              "details": {
                "type": "object",
                "nullable": true,
                "additionalProperties": { "type": "string" }
              }
            },
            "required": ["code", "message"]
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "feedback_count": { "type": "integer" }
        }
      },
      "Round": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "frozen", "archived"] },
          "description": { "type": "string", "nullable": true },
          "feedback_count": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" },
          "starts_at": { "type": "string", "format": "date-time", "nullable": true },
          "ends_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "Webhook": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "url": { "type": "string", "format": "uri" },
          "events": {
            "type": "array",
            "items": { "type": "string", "enum": ["feedback.created", "feedback.updated", "feedback.deleted"] }
          },
          "active": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "paths": {
...
    "/projects": {
      "get": {
        "summary": "List all projects",
        "tags": ["Projects"],
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "maximum": 100 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "projects": { "type": "array", "items": { "$ref": "#/components/schemas/Project" } },
                    "total": { "type": "integer" },
                    "limit": { "type": "integer" },
                    "offset": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/projects/{projectId}": {
      "get": {
        "summary": "Get project details",
        "tags": ["Projects"],
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Project" } } } },
          "404": { "description": "Project not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } } } }
        }
      }
    },
    "/projects/{projectId}/rounds": {
      "get": {
        "summary": "List rounds in project",
        "tags": ["Projects"],
        "parameters": [
          { "name": "projectId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rounds": { "type": "array", "items": { "$ref": "#/components/schemas/Round" } }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/webhooks": {
      "post": {
        "summary": "Register webhook endpoint",
        "tags": ["Webhooks"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url", "events"],
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "events": {
                    "type": "array",
                    "items": { "type": "string", "enum": ["feedback.created", "feedback.updated", "feedback.deleted"] }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/Webhook" },
                    {
                      "type": "object",
                      "properties": {
                        "secret": { "type": "string", "description": "HMAC secret (shown only once)" }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List registered webhooks",
        "tags": ["Webhooks"],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "webhooks": { "type": "array", "items": { "$ref": "#/components/schemas/Webhook" } }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/webhooks/{webhookId}": {
      "patch": {
        "summary": "Update webhook",
        "tags": ["Webhooks"],
        "parameters": [
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url": { "type": "string", "format": "uri" },
                  "events": {
                    "type": "array",
                    "items": { "type": "string", "enum": ["feedback.created", "feedback.updated", "feedback.deleted"] }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Webhook" } } } }
        }
      },
      "delete": {
        "summary": "Deactivate webhook",
        "tags": ["Webhooks"],
        "parameters": [
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Webhook deactivated" }
        }
      }
    },
    "/webhooks/{webhookId}/test": {
      "post": {
        "summary": "Send test event to webhook",
        "tags": ["Webhooks"],
        "parameters": [
          { "name": "webhookId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "statusCode": { "type": "integer", "nullable": true },
                    "error": { "type": "string", "nullable": true }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}